+++ src/postgres.js
@@ -26,10 +26,102 @@ squel.flavours['postgres'] = function(_squel) {
     }
   }
 
+  // WITH
+  cls.WithBlock = class extends cls.Block {
+    constructor (options) {
+      super(options);
+      this._tables = [];
+    }
+
+    with (alias, table) {
+      this._tables.push({alias, table});
+    }
+
+    _toParamString(options = {}) {
+      var parts  = [];
+      var values = [];
+
+      for (let {alias, table} of this._tables) {
+        let ret = table._toParamString({
+          buildParameterized: options.buildParameterized,
+          nested: true
+        });
+
+        parts.push(`${alias} AS ${ret.text}`);
+        values.push(...ret.values);
+      }
+
+      return {
+        text: parts.length ? `WITH ${parts.join(', ')}` : '',
+        values
+      };
+    }
+  }
+
+  // DISTINCT [ON]
+  cls.DistinctOnBlock = class extends cls.Block {
+    constructor(options) {
+      super(options);
+
+      this._distinctFields = [];
+    }
+
+    distinct(...fields) {
+      this._useDistinct = true;
+
+      // Add all fields to the DISTINCT ON clause.
+      fields.forEach((field) => {
+        this._distinctFields.push(this._sanitizeField(field));
+      });
+    }
+
+    _toParamString() {
+      let text = '';
+
+      if (this._useDistinct) {
+        text = 'DISTINCT';
+
+        if (this._distinctFields.length) {
+            text += ` ON (${this._distinctFields.join(', ')})`;
+        }
+      }
+
+      return {
+        text,
+        values: []
+      };
+    }
+  }
+
+  // SELECT query builder.
+  cls.Select = class extends cls.QueryBuilder {
+    constructor (options, blocks = null) {
+      blocks = blocks || [
+        new cls.WithBlock(options),
+        new cls.StringBlock(options, 'SELECT'),
+        new cls.FunctionBlock(options),
+        new cls.DistinctOnBlock(options),
+        new cls.GetFieldBlock(options),
+        new cls.FromTableBlock(options),
+        new cls.JoinBlock(options),
+        new cls.WhereBlock(options),
+        new cls.GroupByBlock(options),
+        new cls.HavingBlock(options),
+        new cls.OrderByBlock(options),
+        new cls.LimitBlock(options),
+        new cls.OffsetBlock(options),
+        new cls.UnionBlock(options)
+      ];
+
+      super(options, blocks);
+    }
+  }
+
   // INSERT query builder
   cls.Insert = class extends cls.QueryBuilder {
     constructor (options, blocks = null) {
       blocks = blocks || [
+        new cls.WithBlock(options),
         new cls.StringBlock(options, 'INSERT'),
         new cls.IntoTableBlock(options),
         new cls.InsertFieldValueBlock(options),
@@ -45,6 +137,7 @@ squel.flavours['postgres'] = function(_squel) {
   cls.Update = class extends cls.QueryBuilder {
     constructor (options, blocks = null) {
       blocks = blocks || [
+        new cls.WithBlock(options),
         new cls.StringBlock(options, 'UPDATE'),
         new cls.UpdateTableBlock(options),
         new cls.SetFieldBlock(options),
@@ -63,6 +156,7 @@ squel.flavours['postgres'] = function(_squel) {
   cls.Delete = class extends cls.QueryBuilder {
     constructor (options, blocks = null) {
       blocks = blocks || [
+        new cls.WithBlock(options),
         new cls.StringBlock(options, 'DELETE'),
         new cls.TargetTableBlock(options),
         new cls.FromTableBlock(_extend({}, options, {
@@ -79,4 +173,3 @@ squel.flavours['postgres'] = function(_squel) {
     }
   }
 }
-